src/PrototypeServer/requirements.txt
Flask==1.1.1
numpy==1.16.4
Pillow==6.1.0
